% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.test.R
\name{bootstrap.test}
\alias{bootstrap.test}
\title{Bootstrap test}
\usage{
bootstrap.test(distribution, subset, statistic = mean, replicates = 100,
  alternative = "two-sided", ...)
}
\arguments{
\item{distribution}{A \code{numeric} distribution to draw from.}

\item{subset}{A \code{numeric} vector of the elements of the distribution to test.}

\item{statistic}{The statistic to measure from the distribution (a \code{function}).}

\item{replicates}{A \code{numeric} value for the number of replicates (\code{default = 100}).}

\item{alternative}{Optional, if the parameter is tested, what is the alternative hypothesis. Can be \code{"two-sided"} (default), \code{"greater"} or \code{"lesser"}.}

\item{...}{Any optional arguments to be passed to \code{test}.}
}
\value{
This function returns a \code{"randtest"} object that can be passed to the generic S3 functions \code{\link[randtest]{print.randtest}} or \code{\link[randtest]{plot.randtest}}.
The output also contains to extra elements \code{output$observed} and \code{output$random} containing the raw results of respectively the observed and random tests.
}
\description{
Performs a bootstrap test
}
\details{
First, the subset of the distribution is compared to the whole distribution (observed difference).
Second, random equally sized subsets are compared to the whole distribution (random differences).
If the observed difference falls out of the random differences distribution, the differences are significant.

If \code{test.parameter} is set to \code{TRUE}, the code tests whether the resulting test parameters between the observed subset and the random ones are significantly different (base on the same procedure as in \code{link[ade4]{rantest}}).
}
\examples{
## Loading the geomorph dataset
require(geomorph)
data(plethodon)

## Performing the Procrustes superimposition
proc_super <- gpagen(plethodon$land, print.progress = FALSE)

## Getting the two most different specimen based on their landmark change radii
var_range <- variation.range(proc_super)

set.seed(1)

## Selecting 6 random landmarks
random_part <- sample(1:nrow(var_range), 6)

## Testing whether this partition has a different median than expected by change
boot_test <- bootstrap.test(var_range[, "radius"], random_part, statistic = median)

## Summarising the results
boot_test

## Plotting the results
plot(boot_test)

}
\seealso{
\code{link{rand.test}}, \code{\link[ade4]{randtest}}
}
\author{
Thomas Guillerme
}
