% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.area.R
\name{coordinates.area}
\alias{coordinates.area}
\title{Coordinates area}
\usage{
coordinates.area(data, what = "radius")
}
\arguments{
\item{data}{The \code{matrix} of coordinate differences.}

\item{what}{Which element of the coordinate differences to use (can be the \code{numeric} value of the column or the name as \code{character}).}
}
\description{
Calculates the area of the coordinates differences
}
\examples{
## Loading the geomorph dataset
require(geomorph)
data(plethodon)

## Performing the Procrustes superimposition
proc_super <- gpagen(plethodon$land, print.progress = FALSE)

## Getting the spherical coordinates difference between the two first specimen
difference <- coordinates.difference(proc_super$coords[, , 1], proc_super$coords[, , 2],
                                     type = "spherical", angle = "degree")

## Calculate the area of the coordinates differences
coordinates.area(difference[[1]])

}
\seealso{
\code{\link{coordinates.difference}}
}
\author{
Thomas Guillerme
}
