% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.difference.R
\name{coordinates.difference}
\alias{coordinates.difference}
\title{Coordinates differences}
\usage{
coordinates.difference(coordinates, reference, type = "cartesian",
  angle = "radian")
}
\arguments{
\item{coordinates}{An \code{array}, \code{list} or \code{matrix} of coordinates to be compared to the reference.}

\item{reference}{A \code{matrix} of reference coordinates. If missing and \code{coordinates} is an \code{array} or \code{list}, the first element of \code{coordinates} is used.}

\item{type}{the type of coordinates to output: can be \code{"cartesian"} (x0,y0,x1,y1,... format), or  \code{"spherical"} (radius, polar, azimuth).}

\item{angle}{optional, whether display angles in radian (\code{"radian"} - default) or in degrees (\code{"degree"}).}
}
\description{
Calculates coordinate differences (e.g. from procrustes superimpositions)
}
\examples{
## Loading the geomorph dataset
require(geomorph)
data(plethodon)

## Performing the Procrustes superimposition
proc_super <- geomorph::gpagen(plethodon$land, print.progress = FALSE)

## Getting the coordinates differences from the consensus
cartesian_diff <- coordinates.difference(proc_super$coords, proc_super$consensus)

## The coordinates of the differences between the first specimen and the consensus
head(cartesian_diff[[1]])

## Getting the spherical coordinates difference between the two first specimen
coordinates.difference(proc_super$coords[, , 1], proc_super$coords[, , 2],
                       type = "spherical", angle = "degree")

}
\seealso{
\code{\link{variation.range}}
}
\author{
Thomas Guillerme
}
