% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRefToTarget.heat.R
\name{plotRefToTarget.heat}
\alias{plotRefToTarget.heat}
\title{Plot shape differences as PlotRefToTarget with heat}
\usage{
plotRefToTarget.heat(M1, M2, method = "vector",
  plotRefToTarget.args = list(...), ..., col = list("grey", "black"),
  col.val, pt.size)
}
\arguments{
\item{M1}{Matrix of landmark coordinates for the first (reference) specimen}

\item{M2}{Matrix of landmark coordinates for the second (target) specimen}

\item{method}{Method used to visualize shape difference: either \code{"vector"} (default) or \code{"points"}. For \code{"surface"} or \code{"TPS"}, use \code{\link[geomorph]{plotRefToTarget}}.}

\item{plotRefToTarget.args}{additional arguments to be passed to \code{\link[geomorph]{plotRefTarget}}.}

\item{...}{Additional parameters to \code{\link[graphics]{plot}} or \code{\link[rgl]{plot3d}}.}

\item{col}{Either a single color value (\code{"character"}), vector of values or function for colouring both points and vectors; or a list of two of any of these three elements. The first argument is passed to the points and the second to the vectors (\code{default = list("grey", "black")} for grey points and black vectors - see details).}

\item{col.val}{Optional, if \code{col} is a function for assigning colours (or a list of two functions), which values to pass to the function to assign the colors).}

\item{pt.size}{Optional, the size of the points (\code{default = 1}). This argument can overwrite \code{gridPars$pt.size} passed to \code{...}.}
}
\description{
Modification of PlotRefToTarget to allow to add heat maps (only works for \code{methods = c("points", "vector")})
}
\details{
\code{col} arguments are passed to \code{\link[graphics]{plot}} or \code{\link[rgl]{plot3d}} as follows:
\itemize{
   \item if col is a single value or a single vector, it passes it to both points and colors arguments;
   \item if col is a single function, it applies it to both points and colors arguments based on the argument \code{col.val};
   \item if col is a list of two values, vectors or functions, it will apply the first element to the points and the second to the vectors the same way as described above;
}
}
\examples{
## Loading the geomorph dataset
require(geomorph)
data(plethodon)

## Performing the Procrustes superimposition
proc_super <- gpagen(plethodon$land, print.progress = FALSE)

## Getting the range of variation
variation <- variation.range(proc_super, return.ID = TRUE)

## Selecting the coordinates and the variation vector
M1 <- proc_super$coords[, , variation$min.max[1]]
M2 <- proc_super$coords[, , variation$min.max[2]]
var_val <- variation$range[, 1]
 
## Plot the variation
plotRefToTarget.heat(M1, M2, method = "vector")

## A weird looking plethodon
plotRefToTarget.heat(M1, M2, method = "points", col = list(rainbow, "pink"), col.val = var_val,
                pt.size = 2.5, plotRefToTarget.args = list(mag = 3, outline = plethodon$outline))

\dontrun{
## Loading the scallops 3D data from geomorph
require(geomorph)
data(scallops)

## Procrustes superimposition
procrustes <- gpagen(scallops$coorddata)

## Getting the range of variation
variation <- variation.range(procrustes, return.ID = TRUE)

## Selecting the coordinates and the variation vector
M1 <- procrustes$coords[, , variation$min.max[1]]
M2 <- procrustes$coords[, , variation$min.max[2]]
var_val <- variation$range[, 1]
 
## Plot the variation in 3D
plotRefToTarget.heat(M1, M2, method = "vector",  col.val = var_val,
                     col = list(grDevices::heat.colors, "grey"))
}

}
\seealso{
\code{\link{variation.range}}
}
\author{
Dean Adams, Emma Sherratt & Michael Collyer - modified by Thomas Guillerme
}
