% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.test.R
\name{rand.test}
\alias{rand.test}
\title{Random test}
\usage{
rand.test(distribution, subset, test, replicates = 100, resample = TRUE,
  rarefaction, test.parameter = FALSE, parameter, alternative = "two-sided",
  ...)
}
\arguments{
\item{distribution}{A \code{numeric} distribution to draw from.}

\item{subset}{A \code{numeric} vector of the elements of the distribution to test.}

\item{test}{The test to apply (a \code{function}).}

\item{replicates}{A \code{numeric} value for the number of replicates (\code{default = 100}).}

\item{resample}{\code{logical} wether to resample the full distribution (\code{TRUE}) or the distribution without the subset (\code{FALSE}).}

\item{rarefaction}{Optional, a \code{numeric} value for rarefying the subset.}

\item{test.parameter}{Optional, whether to test the calculated parameter (\code{TRUE}) or not (\code{FALSE} - default).}

\item{parameter}{Optional,  if the parameter is tested, which parameter to select (can be left empty if \code{test} outputs a single value or a named object containing a \code{statistic} element).}

\item{alternative}{Optional, if the parameter is tested, what is the alternative hypothesis. Can be \code{"two-sided"} (default), \code{"greater"} or \code{"lesser"}.}

\item{...}{Any optional arguments to be passed to \code{test}.}
}
\value{
This function returns a \code{"randtest"} object that can be passed to the generic S3 functions \code{\link[randtest]{print.randtest}} or \code{\link[randtest]{plot.randtest}}.
The output also contains to extra elements \code{output$observed} and \code{output$random} containing the raw results of respectively the observed and random tests.
}
\description{
Performs a random test
}
\details{
First, the subset of the distribution is compared to the whole distribution (observed difference).
Second, random equally sized subsets are compared to the whole distribution (random differences).
If the observed difference falls out of the random differences distribution, the differences are significant.

If \code{test.parameter} is set to \code{TRUE}, the code tests whether the resulting test parameters between the observed subset and the random ones are significantly different (base on the same procedure as in \code{link[ade4]{rantest}}).
}
\examples{
## Loading the geomorph dataset
require(geomorph)
data(plethodon)

## Performing the Procrustes superimposition
proc_super <- gpagen(plethodon$land, print.progress = FALSE)

## Getting the two most different specimen based on their landmark change radii
var_range <- variation.range(proc_super)

set.seed(1)

## Selecting 6 random landmarks
random_part <- sample(1:nrow(var_range), 6)

## Testing the difference between the two sets of landmarks
stats::t.test(var_range[random_part, "radius"], var_range[-random_part, "radius"])

## Testing whether this difference is expected by chance
random_test <- rand.test(var_range[, "radius"], random_part, test = stats::t.test,
                        test.parameter = TRUE)

## Summarising the results
random_test

## Plotting the results
plot(random_test)

## Rarefying the area difference to 4 elements without testing the parameter
rarefy_test <- rand.test(var_range[, "radius"], random_part, rarefaction = 5,
                         test = t.test)
plot(rarefy_test)

}
\seealso{
\code{link[ade4]{randtest}}, \code{\link{bootstrap.test}}
}
\author{
Thomas Guillerme
}
