% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.difference.R
\name{coordinates.difference}
\alias{coordinates.difference}
\title{Coordinates differences}
\usage{
coordinates.difference(coordinates, reference, type = "cartesian",
  angle = "radian", absolute.distance = TRUE)
}
\arguments{
\item{coordinates}{An \code{array}, \code{list} or \code{matrix} of coordinates to be compared to the reference.}

\item{reference}{A \code{matrix} of reference coordinates. If missing and \code{coordinates} is an \code{array} or \code{list}, the first element of \code{coordinates} is used.}

\item{type}{the type of coordinates to output: can be \code{"cartesian"} (x0,y0,x1,y1,... format), \code{"spherical"} (radius, polar, azimuth) or \code{"vector"} (length, angle(s)).}

\item{angle}{optional, whether display angles in radian (\code{"radian"} - default) or in degrees (\code{"degree"}).}

\item{absolute.distance}{\code{logical}, when using \code{"vector"}, whether to use the absolute distance (from the centroid) or the relative one (from the reference landmark). See details.}
}
\description{
Calculates coordinate differences (e.g. from procrustes superimpositions)
}
\details{
When using \code{type = "vector"} with \code{absolute.distance = TRUE}, the distance between two landmarks A and A' is calculated as d(0,A') - d(0,A) where 0 is the centroid of the shape to analysis.
A positive absolute distance means that A' is further away from the centroid than A.
A negative absolute distance means that A' closer to the centroid than A.
When using \code{absolute.distance = FALSE}, the distance is calculated as d(A,A') and is always positive.
}
\examples{
## Loading the geomorph dataset
require(geomorph)
data(plethodon)

## Performing the Procrustes superimposition
proc_super <- geomorph::gpagen(plethodon$land, print.progress = FALSE)

## Getting the coordinates differences from the consensus
cartesian_diff <- coordinates.difference(proc_super$coords, proc_super$consensus)

## The coordinates of the differences between the first specimen and the consensus
head(cartesian_diff[[1]])

## Getting the spherical coordinates difference between the two first specimen
coordinates.difference(proc_super$coords[, , 1], proc_super$coords[, , 2],
                       type = "spherical", angle = "degree")

## Getting the vector coordinates for the same specimen in relative distance
coordinates.difference(proc_super$coords[, , 1], proc_super$coords[, , 2],
                       type = "vector", angle = "degree", absolute.distance = FALSE)

## Getting the vector same coordinates in absolute distances
coordinates.difference(proc_super$coords[, , 1], proc_super$coords[, , 2],
                       type = "vector", angle = "degree")
}
\seealso{
\code{\link{variation.range}}
}
\author{
Thomas Guillerme
}
