% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation.range.R
\name{variation.range}
\alias{variation.range}
\title{Range of variation}
\usage{
variation.range(procrustes, type = "spherical", angle = "degree",
  what = "radius", ordination, axis, return.ID = FALSE, CI,
  CI.hdr = FALSE, CI.pre = FALSE)
}
\arguments{
\item{procrustes}{Procrustes data of class \code{"gpagen"}.}

\item{type}{Which type of coordinates to calculate (see \code{\link{coordinates.differences}} - default is \code{"sperical"}). See details.}

\item{angle}{Which type of angle to calculate (see \code{\link{coordinates.differences}} - default is \code{"degree"}).}

\item{what}{Which element from the \code{\link{coordinates.differences}} to use (default is \code{"radius"}).}

\item{ordination}{Optional, either \code{TRUE} to perform an ordination or directly an ordinated PCA matrix (\code{"prcomp"}) to calculate the range from there.}

\item{axis}{Optional, if an ordinated matrix is used, which axis (axes) to use. If left empty, all the axes will be used.}

\item{return.ID}{\code{logical}, whether to return the ID of the max/min specimens or not.}

\item{CI}{Optional, a value of confidence interval to use (rather than the max/min).}

\item{CI.hdr}{Logical, whether to use the proper CI estimation from \code{\link[hdrcde]{hdr}} (\code{TRUE}) or not (\code{FALSE} - default) see details.}

\item{CI.pre}{Logical, whether apply the CI before (\code{TRUE}) or during (\code{FALSE} - default) the variation range calculation; see details.}
}
\description{
Selecting the range of differences between the maximum and minimum variation specimen
}
\details{
When \code{CI.hdr = FALSE} the specimen removed are the ones >= CI from the consensus and >= CI from the furthest from the consensus. In other words, the specimen the closest to the CI boundary is first chosen (the max_specimen), the distance are measured from this specimen and then the other specimen the closest to the same CI boundary is chosen (the min_specimen).
When \code{CI.hdr = TRUE} is used, the specimen from both size of the confidence interval are removed straight from the first step (i.e. only the specimens with the CI in terms of distance from the consensus are considered).

When \code{CI.pre = TRUE}, the extreme specimens are removed from the distribution just after calculating the distances from the consensus (i.e. before selecting the maximum and minimum specimen).
}
\examples{
## Loading the geomorph dataset
require(geomorph)
data(plethodon)

## Performing the Procrustes superimposition
proc_super <- gpagen(plethodon$land, print.progress = FALSE)


## Getting the two most different specimen based on their landmark change radii
spec_range <- variation.range(proc_super, return.ID = TRUE)

## The minimum and maximum specimen
spec_range$min.max

## The range of variation per landmark
spec_range$range


## Getting the two most different specimen based on the first axis of the ordination
variation.range(proc_super, ordination = TRUE, axis = 1, type = "vector", what = "length")


## Getting the range variation between specimen using a 95 confidence interval range
spec_range95 <- variation.range(proc_super, CI = 0.95, return.ID = TRUE)

## The absolute maximum and minimum specimens
spec_range$min.max

## The lower and upper 95\% range CI specimens
spec_range95$min.max

}
\seealso{
\code{\link{coordinates.difference}}, \code{\link{area.diff}}, \code{\link{rand.test}}
}
\author{
Thomas Guillerme
}
